Update aLight.vs
The shader now uses the engine's internal map scale because `VS.Client.mapView` is only changed when manually changing it. So alot of misinformation was happening. This new change ensures that the library will have the valid scale of the map at all times. This library is now compatible whether `VS.Client.mapView.preventAutoScale` is enabled or not! WOOOOOT!!