feat: added separate dimension culling
You can now specify the culling distance in each dimension now, as well as the fading distance in each dimension.
This setting has the possibility of being `jarring` if too inconsistent values are used, so try to put the distances outside the `viewing` range of the screen. So it transitions off screen if you are using it in tandem with `fadeDistance`.