`fix`: fixed a dormant issue with the culling system
It was always `culling` no matter if the screen had changed, and this left some lights permanently stuck at a certain value until the screen moved. This was due to the `screenPos` changing multiple times during runtime due to the engine. Forcing the lights to be culled and then unculled when the `screenPos` changed. But the lights got stuck in a cull loop which didn't change their brightness value back to normal.

Fixed a dormant issue with if the only light that existed was the mouse light, then it would not be updated due to it never reaching the `addLightUniforms` function.
Changed the code inside of `uncull` to remove the light from the culled lights array before it adds it to the `lights` array. So that the lights is not in two arrays at once.