`fix`: Removed some redundancies that made useless errors be thrown
Removed the `removing` of the light from the `destroyLight` function and let `this.removeLightUniforms(light)` handle it. This way there will be no errors of not finding the light since it was removed already anymore.

Removed the adding of `culledLights` length to the `lights` length when checking for `MAX_LIGHTS`. To see how it will work out. This will allow more than `MAX_LIGHTS` if they are all not on the screen at once.

Removed checking for the `light`'s owner to remove the light from any array the owner uses to store lights inside of `detachLight`. This is because it gets checked inside of `destroyLight` and its better to do it there.