update: removed a confusing line.
Every time you searched for a light with an id that did not exist, it threw an error to alert you that it did not exist. (console.error('aLight Module [Light ID: \'' + pID + '\']: No %clight', 'font-weight: bold', 'found with that id');)
However this can be a bit confusing to see when you are just trying to do something simple such as 
```js
if (aLight.getLightById('main_lightt')) return true;
```
seeing this error is not really expected. So it has been removed.