maintenance: renamed some variables
Renamed some variables, to open up the namespace for some other libraries.
Added a pAnchor argument to a internal function that assigns the `true` size of the diob. This argument will allow you to change the anchor of the diob in the same call if needed.
Added a comment to better explain what -1 for the `x` and `y` of `cullDistance` does. (It removes the light from the culling system all together)