fix: fixed a issue where anchor wasn't being calculated correctly
Anchor was not being calculated correctly due to a typo. Was doing `width / width` rather than `halfWidth / width`.
Also added a check for `cullDistance` being set to negative 1. Before it would still cull a light if it had a `fadeDistance` set. But a `fadeDistance is useless without culling information. So now it checks if it has culling information as well, and if it doesn't then it will not be culled at all.